rename mapbox contextKey file
