+++ packages/parts/src/routes/5-maps/mapbox.svx
@@ -1,6 +1,5 @@
 <script lang="ts">
   import { Story } from 'kitbook';
-  // import Mapbox from '$lib/maps/Mapbox.svelte';
   import Map from '$lib/maps/mapbox/map/Map.svelte';
 </script>
